requirements.txt
Django~=3.0.1
django-taggit==1.1.0